backend/requirements.txt
pdf2image==1.17.0
pillow==11.2.1
python-pptx==1.0.2
beautifulsoup4==4.13.4